fix(erg): rollback handling of blocks with restored addresses
Issue was resulting from ew relying using supply age to detect new addresses from existing ones. However, when an existing address gets spent entirely in a tx and receives assets with a subsequent tx in the same block, its mean age timestamp will be as if it was a brand new address.
To fix this, the erg worker now marks new addresses explicitly in its output batch.